backend/requirements.txt
fastapi==0.116.1
openai==1.107.2
python-dotenv==1.1.1
requests==2.32.5
supabase==2.23.0
uvicorn==0.35.0
